% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataContext.R
\docType{class}
\name{DataContext-class}
\alias{DataContext-class}
\alias{DataContext}
\title{Classe para acesso ao banco de dados
Manipulação de transações
Conexão
Execução de instrução SQL}
\description{
Classe para acesso ao banco de dados
Manipulação de transações
Conexão
Execução de instrução SQL
}

