% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataContextFactory.R
\docType{class}
\name{DataContextFactory-class}
\alias{DataContextFactory-class}
\alias{DataContextFactory}
\title{Fabricador da classe PDO para acessar aos drivers de banco de dados}
\description{
Fabricador da classe PDO para acessar aos drivers de banco de dados
}

