% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefaultQueryBuilder.R
\docType{class}
\name{DefaultQueryBuilder-class}
\alias{DefaultQueryBuilder}
\alias{DefaultQueryBuilder-class}
\title{Classe para manipulação de String SQL genérica}
\description{
Classe para manipulação de String SQL genérica
}

