% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObjectContext.R
\docType{class}
\name{ObjectContext-class}
\alias{ObjectContext}
\alias{ObjectContext-class}
\title{Classe para manipulação de objetos
- Obter propriedades
- Obter valores
- Obter objeto(s) populados}
\description{
Classe para manipulação de objetos
- Obter propriedades
- Obter valores
- Obter objeto(s) populados
}

