% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryBuilder.R
\docType{class}
\name{QueryBuilder-class}
\alias{QueryBuilder-class}
\alias{QueryBuilder}
\title{Classe para manipulação de String SQL para a função SELECT}
\description{
Classe para manipulação de String SQL para a função SELECT
}

