% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectQueryBuilder.R
\docType{class}
\name{SelectQueryBuilder-class}
\alias{SelectQueryBuilder}
\alias{SelectQueryBuilder-class}
\title{Classe para manipulação de String SQL para a função SELECT}
\description{
Classe para manipulação de String SQL para a função SELECT
}

