% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SqlStatement.R
\docType{class}
\name{SqlStatement-class}
\alias{SqlStatement}
\alias{SqlStatement-class}
\title{Classe para manipulação de String SQL com métodos comuns
- Obter valores das propriedades pra utilizar no SQL}
\description{
Classe para manipulação de String SQL com métodos comuns
- Obter valores das propriedades pra utilizar no SQL
}

