% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SqlStatementDelete.R
\docType{class}
\name{SqlStatementDelete-class}
\alias{SqlStatementDelete}
\alias{SqlStatementDelete-class}
\title{Classe para manipulação de String SQL para a função DELETE}
\description{
Classe para manipulação de String SQL para a função DELETE
}

