% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SqlStatementInsert.R
\docType{class}
\name{SqlStatementInsert-class}
\alias{SqlStatementInsert}
\alias{SqlStatementInsert-class}
\title{Classe para manipulação de String SQL para a função INSERT}
\description{
Classe para manipulação de String SQL para a função INSERT
}

