% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SqlStatementSelect.R
\docType{class}
\name{SqlStatementSelect-class}
\alias{SqlStatementSelect-class}
\alias{SqlStatementSelect}
\title{Classe para manipulação de String SQL para a função SELECT}
\description{
Classe para manipulação de String SQL para a função SELECT
}

