% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SqlStatementUpdate.R
\docType{class}
\name{SqlStatementUpdate-class}
\alias{SqlStatementUpdate}
\alias{SqlStatementUpdate-class}
\title{Classe para manipulação de String SQL para a função UPDATE}
\description{
Classe para manipulação de String SQL para a função UPDATE
}

